CuseRents-backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0
djangorestframework-simplejwt==5.3.1
django-allauth==0.61.1
dj-rest-auth==5.0.2
cloudinary==1.38.0
django-cloudinary-storage==0.3.0
Pillow==10.2.0
stripe==8.2.0
googlemaps==4.10.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
ipython==8.21.0
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0